+++ blinksy/src/driver/clocked/mod.rs
@@ -17,19 +17,22 @@
 //! ## Traits
 //!
 //! - [`ClockedLed`]: Trait defining protocol specifics for a clocked LED chipset
-//! - [`ClockedLedAsync`]: Trait defining protocol specifics for a clocked LED chipset, asynchronously
 //! - [`ClockedWriter`]: Trait for how to write data for a clocked protocol
 //! - [`ClockedWriterAsync`]: Trait for how to write data for a clocked protocol, asynchronously
 //!
 //! ## Drivers
 //!
+//! - [`ClockedDriver`]: Generic driver for clocked LEDs and writers.
 //! - [`ClockedDelayDriver`]: Driver using GPIO bit-banging with a delay timer
 //! - [`ClockedSpiDriver`]: (Recommended) Driver using a hardware SPI peripheral
 //!
 //! ## Example
 //!
 //! ```rust
-//! use blinksy::{color::{ColorCorrection, FromColor, LedRgb, LinearSrgb}, driver::{ClockedLed, ClockedWriter}};
+//! use blinksy::{
+//!     color::{ColorCorrection, FromColor, LedRgb, LinearSrgb},
+//!     driver::{ClockedLed, ClockedWriter},
+//! };
 //!
 //! // Define a new LED chipset with specific protocol requirements
 //! struct MyLed;
@@ -38,31 +41,35 @@
 //!     type Word = u8;
 //!     type Color = LinearSrgb;
 //!
-//!     fn start<W: ClockedWriter<Word = Self::Word>>(writer: &mut W) -> Result<(), W::Error> {
-//!         // Write start frame
-//!         writer.write(&[0x00, 0x00, 0x00, 0x00])
+//!     fn start() -> impl IntoIterator<Item = Self::Word> {
+//!         // Start frame
+//!         [0x00, 0x00, 0x00, 0x00]
 //!     }
 //!
-//!     fn color<W: ClockedWriter<Word = Self::Word>>(
-//!         writer: &mut W,
+//!     fn led(
 //!         color: Self::Color,
 //!         brightness: f32,
 //!         correction: ColorCorrection,
-//!     ) -> Result<(), W::Error> {
-//!         // Write color data for one LED
+//!     ) -> impl IntoIterator<Item = Self::Word> {
+//!         // Color data for one LED
 //!         let linear_srgb = LinearSrgb::from_color(color);
 //!         let rgb = LedRgb::from_linear_srgb(linear_srgb, brightness, correction);
-//!         writer.write(&[0x80, rgb[0], rgb[1], rgb[2]])
+//!         [0x80, rgb[0], rgb[1], rgb[2]]
 //!     }
 //!
-//!     fn end<W: ClockedWriter<Word = Self::Word>>(writer: &mut W, _: usize) -> Result<(), W::Error> {
-//!         // Write end frame
-//!         writer.write(&[0xFF, 0xFF, 0xFF, 0xFF])
+//!     fn end(_: usize) -> impl IntoIterator<Item = Self::Word> {
+//!         // End frame
+//!         [0xFF, 0xFF, 0xFF, 0xFF]
 //!     }
 //! }
 //! ```
 
+use core::marker::PhantomData;
+
 use crate::color::{ColorCorrection, FromColor};
+use crate::driver::Driver;
+#[cfg(feature = "async")]
+use crate::driver::DriverAsync;
 
 mod delay;
 mod spi;
@@ -81,16 +88,47 @@ pub trait ClockedWriter {
     /// The error type that may be returned by write operations.
     type Error;
 
-    /// Writes a slice of words to the protocol.
+    /// Writes an iterator of words to the protocol.
     ///
     /// # Arguments
     ///
-    /// * `words` - Slice of words to write
+    /// * `words` - Iterator of words to write
     ///
     /// # Returns
     ///
     /// Ok(()) on success or an error if the write fails
-    fn write(&mut self, words: &[Self::Word]) -> Result<(), Self::Error>;
+    fn write<Words>(&mut self, words: Words) -> Result<(), Self::Error>
+    where
+        Words: IntoIterator<Item = Self::Word>;
+}
+
+#[cfg(feature = "async")]
+/// Async trait for types that can write data words to a clocked protocol.
+///
+/// This trait abstracts over different implementation methods for writing data
+/// to a clocked protocol, such as bit-banging with GPIOs or using hardware SPI.
+pub trait ClockedWriterAsync {
+    /// The word type (typically u8).
+    type Word: Copy + 'static;
+
+    /// The error type that may be returned by write operations.
+    type Error;
+
+    // See note about allow(async_fn_in_trait) in smart-leds-trait:
+    //   https://github.com/smart-leds-rs/smart-leds-trait/blob/faad5eba0f9c9aa80b1dd17e078e4644f11e7ee0/src/lib.rs#L59-L68
+    #[allow(async_fn_in_trait)]
+    /// Writes an iterator of words to the protocol, asynchronously.
+    ///
+    /// # Arguments
+    ///
+    /// * `words` - Iterator of words to write
+    ///
+    /// # Returns
+    ///
+    /// Ok(()) on success or an error if the write fails
+    async fn write<Words>(&mut self, words: Words) -> Result<(), Self::Error>
+    where
+        Words: IntoIterator<Item = Self::Word>;
 }
 
 /// Trait that defines the protocol specifics for a clocked LED chipset.
@@ -109,55 +147,61 @@ pub trait ClockedLed {
     /// The color representation type.
     type Color;
 
-    /// Writes a start frame to begin a transmission.
-    ///
-    /// This typically sends some form of header that identifies the beginning
-    /// of an LED update sequence.
-    ///
-    /// # Arguments
-    ///
-    /// * `writer` - The writer implementation to use
+    /// A start frame to begin a transmission.
     ///
     /// # Returns
     ///
-    /// Ok(()) on success or an error if the write fails
-    fn start<Writer: ClockedWriter<Word = Self::Word>>(
-        writer: &mut Writer,
-    ) -> Result<(), Writer::Error>;
+    /// An iterator of words to write
+    fn start() -> impl IntoIterator<Item = Self::Word>;
 
-    /// Writes a single color frame for one LED.
+    /// A color frame for a single LED.
     ///
     /// # Arguments
     ///
-    /// * `writer` - The writer implementation to use
     /// * `color` - The color to write
     /// * `brightness` - Global brightness scaling factor (0.0 to 1.0)
     /// * `correction` - Color correction factors
     ///
     /// # Returns
     ///
-    /// Ok(()) on success or an error if the write fails
-    fn color<Writer: ClockedWriter<Word = Self::Word>>(
-        writer: &mut Writer,
+    /// An iterator of words to write
+    fn led(
         color: Self::Color,
         brightness: f32,
         correction: ColorCorrection,
-    ) -> Result<(), Writer::Error>;
+    ) -> impl IntoIterator<Item = Self::Word>;
 
-    /// Writes an end frame to conclude a transmission.
+    /// An end frame to conclude a transmission.
     ///
     /// # Arguments
     ///
-    /// * `writer` - The writer implementation to use
     /// * `pixel_count` - The number of LEDs that were written
     ///
     /// # Returns
     ///
-    /// Ok(()) on success or an error if the write fails
-    fn end<Writer: ClockedWriter<Word = Self::Word>>(
-        writer: &mut Writer,
-        pixel_count: usize,
-    ) -> Result<(), Writer::Error>;
+    /// An iterator of words to write
+    fn end(pixel_count: usize) -> impl IntoIterator<Item = Self::Word>;
+}
+
+/// # Type Parameters
+///
+/// * `Led` - The LED protocol implementation (must implement ClockedLed)
+/// * `Writer` - The clocked writer
+#[derive(Debug)]
+pub struct ClockedDriver<Led, Writer> {
+    /// Marker for the LED protocol type
+    led: PhantomData<Led>,
+    /// Writer implementation for the clocked protocol
+    writer: Writer,
+}
+
+impl<Led, Writer> Driver for ClockedDriver<Led, Writer>
+where
+    Led: ClockedLed,
+    Writer: ClockedWriter<Word = Led::Word>,
+{
+    type Error = Writer::Error;
+    type Color = Led::Color;
 
     /// Writes a complete sequence of colors to the LED chain.
     ///
@@ -168,7 +212,6 @@ pub trait ClockedLed {
     ///
     /// # Arguments
     ///
-    /// * `writer` - The writer implementation to use
     /// * `pixels` - Iterator over colors
     /// * `brightness` - Global brightness scaling factor (0.0 to 1.0)
     /// * `correction` - Color correction factors
@@ -176,127 +219,39 @@ pub trait ClockedLed {
     /// # Returns
     ///
     /// Ok(()) on success or an error if any write operation fails
-    fn clocked_write<Writer, I, C>(
-        writer: &mut Writer,
+    fn write<const PIXEL_COUNT: usize, I, C>(
+        &mut self,
         pixels: I,
         brightness: f32,
         correction: ColorCorrection,
     ) -> Result<(), Writer::Error>
     where
-        Writer: ClockedWriter<Word = Self::Word>,
         I: IntoIterator<Item = C>,
-        Self::Color: FromColor<C>,
+        Led::Color: FromColor<C>,
     {
-        Self::start(writer)?;
+        self.writer.write(Led::start())?;
 
         let mut pixel_count = 0;
         for color in pixels.into_iter() {
-            let color = Self::Color::from_color(color);
-            Self::color(writer, color, brightness, correction)?;
+            let color = Led::Color::from_color(color);
+            self.writer.write(Led::led(color, brightness, correction))?;
             pixel_count += 1;
         }
 
-        Self::end(writer, pixel_count)?;
+        self.writer.write(Led::end(pixel_count))?;
+
         Ok(())
     }
 }
 
 #[cfg(feature = "async")]
-/// Async trait for types that can write data words to a clocked protocol.
-///
-/// This trait abstracts over different implementation methods for writing data
-/// to a clocked protocol, such as bit-banging with GPIOs or using hardware SPI.
-pub trait ClockedWriterAsync {
-    /// The word type (typically u8).
-    type Word: Copy + 'static;
-
-    /// The error type that may be returned by write operations.
-    type Error;
-
-    // See note about allow(async_fn_in_trait) in smart-leds-trait:
-    //   https://github.com/smart-leds-rs/smart-leds-trait/blob/faad5eba0f9c9aa80b1dd17e078e4644f11e7ee0/src/lib.rs#L59-L68
-    #[allow(async_fn_in_trait)]
-    /// Writes a slice of words to the protocol, asynchronously.
-    ///
-    /// # Arguments
-    ///
-    /// * `words` - Slice of words to write
-    ///
-    /// # Returns
-    ///
-    /// Ok(()) on success or an error if the write fails
-    async fn write(&mut self, words: &[Self::Word]) -> Result<(), Self::Error>;
-}
-
-#[cfg(feature = "async")]
-// See note about allow(async_fn_in_trait) in smart-leds-trait:
-//   https://github.com/smart-leds-rs/smart-leds-trait/blob/faad5eba0f9c9aa80b1dd17e078e4644f11e7ee0/src/lib.rs#L59-L68
-#[allow(async_fn_in_trait)]
-/// Async trait that defines the protocol specifics for a clocked LED chipset.
-///
-/// Implementors of this trait specify how to generate the protocol-specific
-/// frames for a particular clocked LED chipset.
-///
-/// # Type Parameters
-///
-/// * `Word` - The basic data unit type (typically u8)
-/// * `Color` - The color representation type
-pub trait ClockedLedAsync {
-    /// The word type (typically u8).
-    type Word: Copy + 'static;
-
-    /// The color representation type.
-    type Color;
-
-    /// Writes a start frame to begin a transmission.
-    ///
-    /// This typically sends some form of header that identifies the beginning
-    /// of an LED update sequence.
-    ///
-    /// # Arguments
-    ///
-    /// * `writer` - The writer implementation to use
-    ///
-    /// # Returns
-    ///
-    /// Ok(()) on success or an error if the write fails
-    async fn start<Writer: ClockedWriterAsync<Word = Self::Word>>(
-        writer: &mut Writer,
-    ) -> Result<(), Writer::Error>;
-
-    /// Writes a single color frame for one LED.
-    ///
-    /// # Arguments
-    ///
-    /// * `writer` - The writer implementation to use
-    /// * `color` - The color to write
-    /// * `brightness` - Global brightness scaling factor (0.0 to 1.0)
-    /// * `correction` - Color correction factors
-    ///
-    /// # Returns
-    ///
-    /// Ok(()) on success or an error if the write fails
-    async fn color<Writer: ClockedWriterAsync<Word = Self::Word>>(
-        writer: &mut Writer,
-        color: Self::Color,
-        brightness: f32,
-        correction: ColorCorrection,
-    ) -> Result<(), Writer::Error>;
-
-    /// Writes an end frame to conclude a transmission.
-    ///
-    /// # Arguments
-    ///
-    /// * `writer` - The writer implementation to use
-    /// * `pixel_count` - The number of LEDs that were written
-    ///
-    /// # Returns
-    ///
-    /// Ok(()) on success or an error if the write fails
-    async fn end<Writer: ClockedWriterAsync<Word = Self::Word>>(
-        writer: &mut Writer,
-        pixel_count: usize,
-    ) -> Result<(), Writer::Error>;
+impl<Led, Writer> DriverAsync for ClockedDriver<Led, Writer>
+where
+    Led: ClockedLed,
+    Writer: ClockedWriterAsync<Word = Led::Word>,
+{
+    type Error = Writer::Error;
+    type Color = Led::Color;
 
     /// Writes a complete sequence of colors to the LED chain.
     ///
@@ -307,7 +262,6 @@ pub trait ClockedLedAsync {
     ///
     /// # Arguments
     ///
-    /// * `writer` - The writer implementation to use
     /// * `pixels` - Iterator over colors
     /// * `brightness` - Global brightness scaling factor (0.0 to 1.0)
     /// * `correction` - Color correction factors
@@ -315,27 +269,29 @@ pub trait ClockedLedAsync {
     /// # Returns
     ///
     /// Ok(()) on success or an error if any write operation fails
-    async fn clocked_write<Writer, I, C>(
-        writer: &mut Writer,
+    async fn write<const PIXEL_COUNT: usize, I, C>(
+        &mut self,
         pixels: I,
         brightness: f32,
         correction: ColorCorrection,
     ) -> Result<(), Writer::Error>
     where
-        Writer: ClockedWriterAsync<Word = Self::Word>,
         I: IntoIterator<Item = C>,
-        Self::Color: FromColor<C>,
+        Led::Color: FromColor<C>,
     {
-        Self::start(writer).await?;
+        self.writer.write(Led::start()).await?;
 
         let mut pixel_count = 0;
         for color in pixels.into_iter() {
-            let color = Self::Color::from_color(color);
-            Self::color(writer, color, brightness, correction).await?;
+            let color = Led::Color::from_color(color);
+            self.writer
+                .write(Led::led(color, brightness, correction))
+                .await?;
             pixel_count += 1;
         }
 
-        Self::end(writer, pixel_count).await?;
+        self.writer.write(Led::end(pixel_count)).await?;
+
         Ok(())
     }
 }